[docs] Migrate data-fetch example to typescript
